Modify module modifiers and improve vector implementation.

+++ src/common/constant.rs
@@ -1,2 +1,2 @@
 pub static LANGUAGE: &'static str = "Rust";
-pub static VECTOR_SIZE: u32 = 1024;
+pub static VECTOR_SIZE: usize = 1024;
